Refactor into util class
