--- src/messagestorage/mod.rs
@@ -1,2 +1,0 @@
-mod message_storage;
-pub use self::message_storage::MessageStorage;
