+++ src/orchestrator/mod.rs
@@ -1,4 +1,3 @@
 mod linearorchestrator;
 
 pub use self::linearorchestrator::LinearOrchestrator;
-pub use self::linearorchestrator::Point;
